refactor(audit): extract osv_protocol crate for shared OSV wire types
