+++ Cargo.toml
@@ -85,6 +85,7 @@ nanoid = "0.5.0"
 nix = { version = "0.29.0", default-features = false }
 once_cell = "1.20.3"
 ordermap = "1.0.0"
+osv_protocol = { path = "crates/osv_protocol" }
 parking_lot = "0.12.4"
 pathdiff = "0.2.3"
 pep440_rs = "0.7.3"
